requirements.txt
# Django Libraries
django~=3.1.3
django_rest_framework~=0.1.0
environs~=7.1.0

# Requests
requests~=2.35.0
